Refactor Flow frustrum culling
Since repetition does not have an index map any more, this commit
introduces an alternate mechanism for managing the "indexMap".

The new RangeController supports a "visibleIndexes" filter.  This array
is functionally equivalent to "indexMap" but is implemented with FRB
(bindings).

The Flow now has an internal RangeController called frustrumController.
It takes as input the Flow's content, or the Flow controller's
organizedContent, and produces as output the content plucked with
"visibleIndexes".  The output is fed into the Flow's repetition.

The updateIndexMap algorithm is retained, but renamed
updateVisibleIndexes, and refactored to maniuplate the visibleIndexes
instead of the indexMap.

FRB had a bug in the "filter" observer, so that has been fixed and the
version predicate updated.  This change does require a fresh `npm
install`.